refactor(app): moves home page into own folder
